backend/requirements.txt
Flask
Pillow
Flask-CORS
Flask-SQLAlchemy
python-dotenv